refactor: Разбиение ProfileService на отдельные сервисы
