+++ src/main/java/com/itmentorcommunityplatform/profileservice/controller/InternalProfileController.java
@@ -5,7 +5,7 @@ import com.itmentorcommunityplatform.profileservice.docs.GetProfileByTgUrlDocs;
 import com.itmentorcommunityplatform.profileservice.docs.UpsertInternalProfileDocs;
 import com.itmentorcommunityplatform.profileservice.dto.request.ProfileUpsertInternalRequestDto;
 import com.itmentorcommunityplatform.profileservice.dto.response.ProfileWithTelegramIdResponseDto;
-import com.itmentorcommunityplatform.profileservice.service.ProfileService;
+import com.itmentorcommunityplatform.profileservice.service.InternalProfileService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,13 +16,13 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class InternalProfileController {
 
-    private final ProfileService profileService;
+    private final InternalProfileService internalProfileService;
 
     @PostMapping("/profile")
     @UpsertInternalProfileDocs
     public ResponseEntity<Void> upsertProfile(
             @RequestBody ProfileUpsertInternalRequestDto dto) {
-        boolean isCreated = profileService.upsertProfile(dto);
+        boolean isCreated = internalProfileService.upsertProfile(dto);
         return isCreated
                 ? ResponseEntity.status(HttpStatus.CREATED).build()
                 : ResponseEntity.ok().build();
@@ -33,7 +33,7 @@ public class InternalProfileController {
     public ResponseEntity<ProfileWithTelegramIdResponseDto> getProfileByGitHubUrl(
             @RequestParam("url") String gitHubUrl) {
 
-        ProfileWithTelegramIdResponseDto profile = profileService.getProfileByGitHubUrl(gitHubUrl);
+        ProfileWithTelegramIdResponseDto profile = internalProfileService.getProfileByGitHubUrl(gitHubUrl);
         return ResponseEntity.ok(profile);
     }
 
@@ -42,8 +42,7 @@ public class InternalProfileController {
     public ResponseEntity<ProfileWithTelegramIdResponseDto> getProfileByTgUrl(
             @RequestParam("url") String tgUrl) {
 
-        ProfileWithTelegramIdResponseDto profile = profileService.getProfileByTgUrl(tgUrl);
-
+        ProfileWithTelegramIdResponseDto profile = internalProfileService.getProfileByTgUrl(tgUrl);
         return ResponseEntity.ok(profile);
     }
 }
