+++ src/main/java/com/itmentorcommunityplatform/profileservice/service/ProfileService.java
@@ -1,25 +1,11 @@
 package com.itmentorcommunityplatform.profileservice.service;
 
-import com.itmentorcommunityplatform.profileservice.client.AuthServiceClient;
-import com.itmentorcommunityplatform.profileservice.domain.Achievement;
 import com.itmentorcommunityplatform.profileservice.domain.Profile;
 import com.itmentorcommunityplatform.profileservice.domain.ProfileDetail;
 import com.itmentorcommunityplatform.profileservice.domain.type.ProfileDetailType;
-import com.itmentorcommunityplatform.profileservice.domain.type.Role;
 import com.itmentorcommunityplatform.profileservice.dto.event.ProjectCreatedEvent;
 import com.itmentorcommunityplatform.profileservice.dto.event.UserCreatedEvent;
-import com.itmentorcommunityplatform.profileservice.dto.request.ProfileUpdateRequestDto;
-import com.itmentorcommunityplatform.profileservice.dto.request.ProfileUpsertInternalRequestDto;
-import com.itmentorcommunityplatform.profileservice.dto.external.UserWithRolesResponseDto;
-import com.itmentorcommunityplatform.profileservice.dto.response.*;
-import com.itmentorcommunityplatform.profileservice.mapper.ProfileMapper;
-import com.itmentorcommunityplatform.profileservice.metrics.ProfileMetrics;
-import com.itmentorcommunityplatform.profileservice.repository.AchievementRepository;
 import com.itmentorcommunityplatform.profileservice.repository.ProfileRepository;
-import com.itmentorcommunityplatform.profileservice.validator.base.BaseProfileDetailValidator;
-import com.itmentorcommunityplatform.profileservice.validator.impl.GithubProfileUrlValidator;
-import com.itmentorcommunityplatform.profileservice.validator.impl.TelegramProfileUrlValidator;
-import com.itmentorcommunityplatform.profileservice.validator.registry.ProfileDetailValidatorRegistry;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -27,8 +13,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
+
+import static com.itmentorcommunityplatform.profileservice.service.ProfileHelperService.mergeProfileDetails;
 
 @Slf4j
 @Service
@@ -36,15 +27,6 @@ import java.util.stream.Collectors;
 public class ProfileService {
 
     private final ProfileRepository profileRepository;
-    private final AchievementRepository achievementRepository;
-    private final ProfileMetrics profileMetrics;
-    private final BaseProfileDetailValidator baseDetailValidator;
-    private final ProfileDetailValidatorRegistry detailValidatorRegistry;
-    private final GithubProfileUrlValidator githubProfileUrlValidator;
-    private final ProfileMapper profileMapper;
-    private final AuthServiceClient authServiceClient;
-    private final TelegramProfileUrlValidator telegramProfileUrlValidator;
-
 
     @Transactional
     public void createOrUpdateProfile(UserCreatedEvent event) {
@@ -93,105 +75,6 @@ public class ProfileService {
         }
     }
 
-
-    public ProfileWithAchievementsResponseDto getCurrentUserProfile(Long telegramUserId) {
-        return profileMetrics.getGetProfileTimer().record(() -> {
-            log.info("Fetching profile for telegramUserId: {}", telegramUserId);
-            try {
-                Profile profile = profileRepository.findByTelegramUserId(telegramUserId)
-                        .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Profile not found"));
-
-                List<Achievement> achievements = achievementRepository.findAllByProfileIdAndPubliclyVisibleTrue(profile.getId());
-
-                profileMetrics.getGetProfileSuccessCounter().increment();
-
-                return profileMapper.mapToProfileWithAchievementsDto(profile.getDetails(), achievements);
-
-            } catch (Exception e) {
-                profileMetrics.getGetProfileErrorCounter().increment();
-                throw e;
-            }
-        });
-    }
-
-    @Transactional
-    public ProfileWithAchievementsResponseDto updateCurrentProfile(Long telegramUserId, ProfileUpdateRequestDto dto, String telegramUsername) {
-        return profileMetrics.getGetProfileTimer().record(() -> {
-            try {
-                Map<String, String> newDetailsMap = dto.getDetails();
-                validateDetails(newDetailsMap);
-
-                if (telegramUsername != null && !telegramUsername.isBlank()) {
-                    newDetailsMap.put("telegram_url", "https://t.me/" + telegramUsername);
-                }
-
-                Profile profile = profileRepository.findByTelegramUserId(telegramUserId)
-                        .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Profile not found"));
-
-                List<Achievement> achievements = achievementRepository.findAllByProfileIdAndPubliclyVisibleTrue(profile.getId());
-
-                Set<ProfileDetail> mergedDetails = mergeProfileDetails(profile.getDetails(), newDetailsMap);
-                profile.setDetails(mergedDetails);
-
-                profileRepository.save(profile);
-                profileMetrics.getGetProfileSuccessCounter().increment();
-
-                return profileMapper.mapToProfileWithAchievementsDto(profile.getDetails(), achievements);
-
-            } catch (Exception e) {
-                profileMetrics.getGetProfileErrorCounter().increment();
-                throw e;
-            }
-        });
-    }
-
-    @Transactional
-    public boolean upsertProfile(ProfileUpsertInternalRequestDto dto) {
-
-        Long telegramUserId = dto.telegramUserId();
-        if (telegramUserId == null) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "'telegram_user_id' must be provided");
-        }
-
-        if (dto.details() == null) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "'details' must be provided");
-        }
-        Map<String, String> newDetailsMap = dto.details().getMap();
-        validateDetails(newDetailsMap);
-
-        Optional<Profile> foundProfile = profileRepository.findByTelegramUserId(telegramUserId);
-        boolean isNewProfile = foundProfile.isEmpty();
-
-        Set<ProfileDetail> existingDetails = foundProfile.map(Profile::getDetails).orElse(Collections.emptySet());
-
-        Set<ProfileDetail> mergedDetails = mergeProfileDetails(existingDetails, newDetailsMap);
-
-        Profile profile = Profile.builder()
-                .id(foundProfile.map(Profile::getId).orElse(null))
-                .telegramUserId(telegramUserId)
-                .details(mergedDetails)
-                .build();
-
-        profileRepository.save(profile);
-
-        return isNewProfile;
-    }
-
-    public ProfileWithTelegramIdResponseDto getProfileByGitHubUrl(String gitHubUrl) {
-
-        githubProfileUrlValidator.validate(gitHubUrl);
-
-        log.info("Searching profile by GitHub URL: {}", gitHubUrl);
-
-        Profile profile = profileRepository.findProfileByGitHubUrl(gitHubUrl)
-                .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND,
-                        "Profile with URL: %s not found".formatted(gitHubUrl)
-                ));
-
-        return profileMapper.mapToProfileWithTelegramIdDto(profile.getTelegramUserId(), profile.getDetails());
-    }
-
     @Transactional(readOnly = true)
     public Optional<Profile> getProfileForEvent(ProjectCreatedEvent event) {
         if (event == null || event.getAuthorTelegramUserId() == null) {
@@ -211,19 +94,6 @@ public class ProfileService {
         return maybeProfile;
     }
 
-    public ProfileWithAchievementsResponseDto getUserProfile(Long profileId) {
-
-        Profile profile = profileRepository.findById(profileId)
-                .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND,
-                        "Profile with id: %s not found".formatted(profileId)
-                ));
-
-        List<Achievement> achievements = achievementRepository.findAllByProfileIdAndPubliclyVisibleTrue(profileId);
-
-        return profileMapper.mapToProfileWithAchievementsDto(profile.getDetails(), achievements);
-    }
-
     @Transactional(readOnly = true)
     public Profile getProfileByTelegramIdOrThrow(Long telegramUserId) {
         return profileRepository.findByTelegramUserId(telegramUserId)
@@ -232,135 +102,5 @@ public class ProfileService {
                     return new ResponseStatusException(HttpStatus.NOT_FOUND,
                             "Profile with Telegram-User-Id %s does not exist".formatted(telegramUserId));
                 });
-    }
-
-    public AllProfilesPaginatedResponseDto getAllProfiles(int pageSize,
-                                                          int pageNumber,
-                                                          Map<String, String> detailFilters) {
-
-        List<Profile> allProfilesPaginated;
-        Long allProfilesCount;
-        int totalPageCount;
-
-        int offset = (pageNumber - 1) * pageSize;
-
-        if (detailFilters == null || detailFilters.isEmpty()) {
-            allProfilesPaginated = profileRepository.findAll(pageSize, offset);
-
-            allProfilesCount = profileRepository.count();
-
-        } else {
-
-            validateDetails(detailFilters);
-
-            List<String[]> detailFiltersList = detailFilters.entrySet()
-                    .stream().map(e -> new String[]{e.getKey().toLowerCase(), e.getValue().toLowerCase()})
-                    .toList();
-
-            allProfilesPaginated = profileRepository.findByDetails(detailFiltersList,
-                    detailFiltersList.size(),
-                    pageSize,
-                    offset);
-
-            allProfilesCount = profileRepository.countFiltered(detailFiltersList, detailFiltersList.size());
-
-        }
-
-        totalPageCount = Math.max((int) Math.ceil((double) allProfilesCount / pageSize), 1);
-
-        if (pageNumber > totalPageCount) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Page number is greater than total page count");
-        }
-
-        List<ProfileWithRolesResponseDto> items;
-
-        try {
-            items = enrichProfilesWithRoles(allProfilesPaginated);
-        } catch (Exception ex) {
-            log.error("All attempts to fetch user roles failed. Reason: {}", ex.getMessage());
-            throw ex;
-        }
-
-        return new AllProfilesPaginatedResponseDto(allProfilesCount, totalPageCount, allProfilesPaginated.size(), pageNumber, items);
-    }
-
-    public ProfileWithTelegramIdResponseDto getProfileByTgUrl(String tgUrl) {
-
-        telegramProfileUrlValidator.validate(tgUrl);
-
-        log.info("Searching profile by Telegram URL: {}", tgUrl);
-
-        Profile profile = profileRepository.findProfileByTgUrl(tgUrl).orElseThrow(
-                () -> new ResponseStatusException(HttpStatus.NOT_FOUND,
-                        "Profile with given Telegram URL not found"));
-
-        return profileMapper.mapToProfileWithTelegramIdDto(profile.getTelegramUserId(), profile.getDetails());
-    }
-
-    private List<ProfileWithRolesResponseDto> enrichProfilesWithRoles(List<Profile> profiles) {
-        if (profiles == null || profiles.isEmpty()) {
-            return List.of();
-        }
-
-        List<Long> telegramIds = profiles.stream()
-                .map(Profile::getTelegramUserId)
-                .toList();
-
-        List<UserWithRolesResponseDto> userRoles = authServiceClient.getAllUsers(telegramIds);
-
-        Map<Long, List<Role>> rolesById = userRoles.stream()
-                .collect(Collectors.toMap(
-                        UserWithRolesResponseDto::telegramUserId,
-                        user -> user.roles().stream()
-                                .map(role -> Role.valueOf(role.toUpperCase()))
-                                .toList()
-                ));
-
-        return profiles.stream()
-                .map(profile -> new ProfileWithRolesResponseDto(
-                        profile.getId(),
-                        new ProfileDetailsResponseDto(profile.getDetails().stream()
-                                .collect(Collectors.toMap(
-                                        ProfileDetail::getDetailName,
-                                        ProfileDetail::getDetailValue
-                                ))),
-                        rolesById.getOrDefault(profile.getTelegramUserId(), List.of())
-                )).toList();
-    }
-
-    private static Set<ProfileDetail> mergeProfileDetails(
-            Set<ProfileDetail> existingDetails,
-            Map<String, String> newDetailsMap
-    ) {
-        Map<String, String> mergedMap = new HashMap<>();
-
-        existingDetails.forEach(detail ->
-                mergedMap.put(detail.getDetailName(), detail.getDetailValue()));
-
-        mergedMap.putAll(newDetailsMap);
-
-        return mergedMap.entrySet().stream()
-                .map(e -> new ProfileDetail(e.getKey(), e.getValue()))
-                .collect(Collectors.toSet());
-    }
-
-    private void validateDetails(Map<String, String> details) {
-        if (details == null || details.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Profile details should not be empty");
-        }
-
-        details.forEach((detailName, detailValue) -> {
-            ProfileDetailType type = ProfileDetailType.fromName(detailName)
-                    .orElseThrow(() -> new ResponseStatusException(
-                            HttpStatus.BAD_REQUEST,
-                            "Unknown detail name: " + detailName
-                    ));
-
-            baseDetailValidator.validate(detailName, detailValue);
-
-            detailValidatorRegistry.getSpecificValidator(type)
-                    .ifPresent(v -> v.validate(detailValue));
-        });
-
     }
 }
